% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add learning function}
\usage{
add(x, y)
}
\arguments{
\item{y}{}
}
\description{
This is a temporary package that will be deleted.
}
\details{
This function adds x and y
}
\examples{
\dontrun{
x = 5
y = 3
add(x, y)
}

}
